Use AddProviderRegistration instead of LoggerFactory.ctor due to version differences
