+++ src/Agent/NewRelic/Agent/Extensions/Providers/Wrapper/MicrosoftExtensionsLogging/Instrumentation.xml
@@ -5,9 +5,9 @@ SPDX-License-Identifier: Apache-2.0
 -->
 <extension xmlns="urn:newrelic-extension">
   <instrumentation>
-    <tracerFactory name="LoggerFactoryWrapper">
+    <tracerFactory name="AddProviderRegistrationWrapper">
       <match assemblyName="Microsoft.Extensions.Logging" className="Microsoft.Extensions.Logging.LoggerFactory">
-        <exactMethodMatcher methodName=".ctor" parameters="System.Collections.Generic.IEnumerable`1[Microsoft.Extensions.Logging.ILoggerProvider],Microsoft.Extensions.Options.IOptionsMonitor`1[Microsoft.Extensions.Logging.LoggerFilterOptions],Microsoft.Extensions.Options.IOptions`1[Microsoft.Extensions.Logging.LoggerFactoryOptions]"/>
+        <exactMethodMatcher methodName="AddProviderRegistration" />
       </match>
     </tracerFactory>
